use uint for len of observer set

+++ x/observer/keeper/msg_server_remove_observer.go
@@ -26,7 +26,7 @@ func (k msgServer) RemoveObserver(
 	// We remove it from both the node account list and the observer set to effectively observing and signing
 	k.RemoveNodeAccount(ctx, msg.ObserverAddress)
 	newCount := k.RemoveObserverFromSet(ctx, msg.ObserverAddress)
-	k.SetLastObserverCount(ctx, &types.LastObserverCount{Count: uint64(newCount)})
+	k.SetLastObserverCount(ctx, &types.LastObserverCount{Count: newCount})
 
 	return &types.MsgRemoveObserverResponse{}, nil
 }
